rename document.[ch]pp to geometry
