+++ src/overlap_checker.cpp
@@ -15,7 +15,7 @@
 #include <cxx_argp_parser.h>
 #include <aixlog.hpp>
 
-#include "document.hpp"
+#include "geometry.hpp"
 #include "utils.hpp"
