+++ src/overlap_collecter.cpp
@@ -10,7 +10,7 @@
 #include <cxx_argp_parser.h>
 #include <aixlog.hpp>
 
-#include "document.hpp"
+#include "geometry.hpp"
 #include "utils.hpp"
 
 static int
