+++ src/step_to_brep.cpp
@@ -35,7 +35,7 @@
 #include <cxx_argp_parser.h>
 #include <aixlog.hpp>
 
-#include "document.hpp"
+#include "geometry.hpp"
 #include "utils.hpp"
 
 static void
